requirements.txt
numpy==2.3.5
pygame==2.6.1
PyQt6==6.10.0
PyQt6-Qt6==6.10.0
PyQt6_sip==13.10.2
pyserial==3.5
